refactor: Make Point3D extend Vector
There is a lot of code that requires a vector from Point3D objects, so it just makes sense to have Point3D extend Vector. This way I can remove code like .getVector() when a Point3D object needs to perform a Vector operation, which just helps to clean up what is already pretty cluttered code.

+++ app/src/main/java/floor/planner/models/Camera.java
@@ -260,21 +260,21 @@ public class Camera {
 
     private void setLookAt(int i) {
         Point3D lookAtPoint = this.pointsOnLookAtCircle.get(i);
-        this.lookAtX = lookAtPoint.getX();
-        this.lookAtY = lookAtPoint.getY();
-        this.lookAtZ = this.pointsOnTiltCircle.get(this.currentZPoint).getZ();
+        this.lookAtX = (float) lookAtPoint.getX();
+        this.lookAtY = (float) lookAtPoint.getY();
+        this.lookAtZ = (float) this.pointsOnTiltCircle.get(this.currentZPoint).getZ();
     }
 
     public void lookUp() {
         this.changeCurrentZPoint(-1);
-        this.lookAtX = this.pointsOnTiltCircle.get(this.currentZPoint).getX();
-        this.lookAtZ = this.pointsOnTiltCircle.get(this.currentZPoint).getZ();
+        this.lookAtX = (float) this.pointsOnTiltCircle.get(this.currentZPoint).getX();
+        this.lookAtZ = (float) this.pointsOnTiltCircle.get(this.currentZPoint).getZ();
     }
 
     public void lookDown() {
         this.changeCurrentZPoint(1);
-        this.lookAtX = this.pointsOnTiltCircle.get(this.currentZPoint).getX();
-        this.lookAtZ = this.pointsOnTiltCircle.get(this.currentZPoint).getZ();
+        this.lookAtX = (float) this.pointsOnTiltCircle.get(this.currentZPoint).getX();
+        this.lookAtZ = (float) this.pointsOnTiltCircle.get(this.currentZPoint).getZ();
     }
 
     public void moveIn() {
